requirements.txt
bs4
requests
numpy
nltk
sklearn
networkx
Pillow
tweepy